#32 - Fixing references to moved objects
